Add GetFileTypeChoices to PluginCatalog for reusable file dialog filters (fixes #256)
