+++ src/SkyCD.Plugin.Host/PluginCatalog.cs
@@ -1,5 +1,9 @@
+using Avalonia.Controls;
+using Avalonia.Platform.Storage;
 using SkyCD.Plugin.Abstractions.Capabilities;
+using SkyCD.Plugin.Abstractions.Capabilities.FileFormats;
 using SkyCD.Plugin.Runtime.Discovery;
+using System.Linq;
 
 namespace SkyCD.Plugin.Host;
 
@@ -25,5 +29,48 @@ public sealed class PluginCatalog
             .SelectMany(plugin => plugin.Capabilities)
             .OfType<TCapability>()
             .ToList();
+    }
+
+    public IReadOnlyList<FilePickerFileType> GetFileTypeChoices(bool allowRead, bool allowWrite)
+    {
+        var fileFormatCapabilities = _plugins
+            .SelectMany(plugin => plugin.Capabilities)
+            .OfType<IFileFormatPluginCapability>()
+            .ToList();
+
+        var formats = new List<FileFormatDescriptor>();
+
+        foreach (var capability in fileFormatCapabilities)
+        {
+            foreach (var format in capability.SupportedFormats)
+            {
+                var canUse = (allowRead && format.CanRead) || (allowWrite && format.CanWrite);
+                if (canUse)
+                {
+                    formats.Add(format);
+                }
+            }
+        }
+
+        var fileTypeChoices = new List<FilePickerFileType>
+        {
+            new FilePickerFileType("All supported formats")
+        };
+
+        foreach (var format in formats.DistinctBy(f => f.FormatId))
+        {
+            var patterns = format.Extensions.Select(ext => $"*{ext}").ToArray();
+            fileTypeChoices.Add(new FilePickerFileType(format.DisplayName)
+            {
+                Patterns = patterns
+            });
+        }
+
+        fileTypeChoices.Add(new FilePickerFileType("All files")
+        {
+            Patterns = ["*.*"]
+        });
+
+        return fileTypeChoices;
     }
 }
